# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Athena API",
    "version": "1.0.0",
    "description": "REST API for Project Athena - Personal productivity and task management"
  },
  "components": {
    "schemas": {
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Unauthorized"]
          },
          "message": {
            "type": "string",
            "default": "Authentication required"
          }
        },
        "required": ["error"]
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Not found"]
          },
          "message": {
            "type": "string",
            "description": "Resource description"
          }
        },
        "required": ["error"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "additionalProperties": {},
            "description": "Additional error details"
          }
        },
        "required": ["error"]
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Task UUID"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "Task title",
            "example": "Review pull request"
          },
          "description": {
            "type": ["string", "null"],
            "description": "Task description",
            "example": "Review the authentication changes in PR #123"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed", "cancelled"],
            "description": "Task status",
            "example": "pending"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "urgent"],
            "description": "Task priority level",
            "example": "medium"
          },
          "deadline": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Task deadline",
            "example": "2025-01-15T17:00:00Z"
          },
          "estimatedMinutes": {
            "type": ["integer", "null"],
            "exclusiveMinimum": 0,
            "description": "Estimated time to complete in minutes",
            "example": 30
          },
          "projectId": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "Associated project UUID"
          },
          "assigneeId": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "Assigned user UUID"
          },
          "creatorId": {
            "type": "string",
            "format": "uuid",
            "description": "Creator user UUID"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2025-01-09T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2025-01-09T12:00:00Z"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "status",
          "priority",
          "deadline",
          "estimatedMinutes",
          "projectId",
          "assigneeId",
          "creatorId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Validation error"]
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field that failed validation"
                },
                "message": {
                  "type": "string",
                  "description": "Validation error message"
                }
              },
              "required": ["field", "message"]
            }
          }
        },
        "required": ["error", "details"]
      },
...
      "TimeBlockResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TimeBlockWithTasks"
          }
        },
        "required": ["data"],
        "description": "Time block response"
      },
      "TimeBlockWithTasks": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TimeBlock"
          },
          {
            "type": "object",
            "properties": {
              "linkedTasks": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/LinkedTask"
                },
                "description": "Tasks linked to this time block with their positions"
              }
            }
          }
        ]
      },
      "LinkedTask": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Task"
          },
          {
            "type": "object",
            "properties": {
              "position": {
                "type": "integer",
                "description": "Task position within the time block",
                "example": 0
              }
            },
            "required": ["position"]
          }
        ]
      },
      "TimeBlock": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Time block UUID"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "Time block label",
            "example": "Focus Time"
          },
          "description": {
            "type": ["string", "null"],
            "description": "Time block description",
            "example": "Deep work session for coding"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "Block start time",
            "example": "2025-01-10T09:00:00Z"
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "description": "Block end time",
            "example": "2025-01-10T11:00:00Z"
          },
          "color": {
            "type": ["string", "null"],
            "description": "Block color hex code",
            "example": "#3B82F6"
          },
          "recurrenceRule": {
            "type": ["string", "null"],
            "description": "iCalendar RRULE for recurring time blocks",
            "example": "FREQ=WEEKLY;BYDAY=MO,WE,FR"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "description": "Owner user UUID"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Soft delete timestamp",
            "example": "2025-01-09T12:00:00Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2025-01-09T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2025-01-09T12:00:00Z"
          }
        },
        "required": [
          "id",
          "label",
          "description",
          "startTime",
          "endTime",
          "color",
          "recurrenceRule",
          "ownerId",
          "deletedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTimeBlockRequest": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "Time block label",
            "example": "Focus Time"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "description": "Time block description"
          },
          "startTime": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Block start time (ISO 8601)",
            "example": "2025-01-10T09:00:00Z"
          },
          "endTime": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Block end time (ISO 8601)",
            "example": "2025-01-10T11:00:00Z"
          },
          "color": {
            "type": "string",
            "maxLength": 50,
            "description": "Block color hex code",
            "example": "#3B82F6"
          },
          "recurrenceRule": {
            "type": "string",
            "maxLength": 500,
            "description": "iCalendar RRULE for recurring time blocks"
          },
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Task UUIDs to link to this time block"
          }
        },
        "required": ["label", "startTime", "endTime"]
      },
      "LinkTaskRequest": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid",
            "description": "Task UUID to link"
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "description": "Position in the task list (defaults to end)"
          }
        },
        "required": ["taskId"]
      },
      "TimeBlockListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeBlockWithTasks"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of items"
          }
        },
        "required": ["data"],
        "description": "Time block list response"
      },
      "ReorderTasksRequest": {
        "type": "object",
        "properties": {
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Task UUIDs in desired order"
          }
        },
        "required": ["taskIds"]
      },
      "UpdateTimeBlockRequest": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "Time block label"
          },
          "description": {
            "type": ["string", "null"],
            "maxLength": 10000,
            "description": "Time block description (null to clear)"
          },
          "startTime": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Block start time (ISO 8601)"
          },
          "endTime": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Block end time (ISO 8601)"
          },
          "color": {
            "type": ["string", "null"],
            "maxLength": 50,
            "description": "Block color (null to clear)"
          },
          "recurrenceRule": {
            "type": ["string", "null"],
            "maxLength": 500,
            "description": "Recurrence rule (null to clear)"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/time-blocks/": {
      "post": {
        "tags": ["Time Blocks"],
        "summary": "Create a time block",
        "description": "Create a new time block.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimeBlockRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Time block created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeBlockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Time Blocks"],
        "summary": "List time blocks",
        "description": "Retrieve a list of time blocks with optional date filtering and pagination.",
        "parameters": [
          {
            "schema": {
              "type": ["string", "null"],
              "format": "date-time",
              "description": "Filter time blocks starting on or after this date",
              "example": "2025-01-01T00:00:00Z"
            },
            "required": false,
            "description": "Filter time blocks starting on or after this date",
            "name": "startDate",
            "in": "query"
          },
          {
            "schema": {
              "type": ["string", "null"],
              "format": "date-time",
              "description": "Filter time blocks ending before this date",
              "example": "2025-01-31T23:59:59Z"
            },
            "required": false,
            "description": "Filter time blocks ending before this date",
            "name": "endDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "description": "Maximum number of time blocks to return",
              "example": 50
            },
            "required": false,
            "description": "Maximum number of time blocks to return",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of time blocks to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of time blocks to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Time blocks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeBlockListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/api/time-blocks/{id}": {
      "delete": {
        "tags": ["Time Blocks"],
        "summary": "Delete a time block",
        "description": "Soft-delete a time block by its ID.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Time block UUID",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            },
            "required": true,
            "description": "Time block UUID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Time block deleted successfully"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Time block not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Time Blocks"],
        "summary": "Get a time block",
        "description": "Retrieve a single time block by its ID.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Time block UUID",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            },
            "required": true,
            "description": "Time block UUID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Time block retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeBlockResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Time block not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Time Blocks"],
        "summary": "Update a time block",
        "description": "Update an existing time block. Only provided fields will be updated.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Time block UUID",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            },
            "required": true,
            "description": "Time block UUID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTimeBlockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Time block updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeBlockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Time block not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/time-blocks/{id}/tasks": {
      "post": {
        "tags": ["Time Blocks"],
        "summary": "Link task to time block",
        "description": "Link a task to a time block.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Time block UUID",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            },
            "required": true,
            "description": "Time block UUID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task linked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeBlockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or task already linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Time block or task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/time-blocks/{id}/tasks/order": {
      "put": {
        "tags": ["Time Blocks"],
        "summary": "Reorder tasks in time block",
        "description": "Reorder the tasks linked to a time block.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Time block UUID",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            },
            "required": true,
            "description": "Time block UUID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTasksRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tasks reordered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeBlockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (e.g., task IDs do not match)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Time block not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/time-blocks/{id}/tasks/{taskId}": {
      "delete": {
        "tags": ["Time Blocks"],
        "summary": "Unlink task from time block",
        "description": "Remove the link between a task and a time block.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Time block UUID"
            },
            "required": true,
            "description": "Time block UUID",
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Task UUID"
            },
            "required": true,
            "description": "Task UUID",
            "name": "taskId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Task unlinked successfully"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Time block or task link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}